UNTESTED rough idea to tweak rate limits
This is probably not the code we want, but it's the simplest way I
could think of to show my idea.

To let instance users make more requests than non-logged-in users, we
could just set a higher `factor`. In this example, I used 2.

This is equivalent to setting a conditional role, that applies to all
local users, that sets the `rateLimitFactor` to 2 (or, better said, to
twice whatever the admins have set already).

A different approach would be to equip each endpoint with 2 rate
limits, for logged-in users and for others.

Discuss.